requirements.txt
streamlit
google-generativeai
googletrans==4.0.0rc1
python-dotenv
